% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{logger.tester.function}
\alias{logger.tester.function}
\title{Testing logging from package}
\usage{
logger.tester.function(level, msg)
}
\arguments{
\item{level}{foo}

\item{msg}{bar}
}
\description{
Testing logging from package
}
