% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{logger_info_tester_function}
\alias{logger_info_tester_function}
\title{Testing logging INFO from package}
\usage{
logger_info_tester_function(msg)
}
\arguments{
\item{msg}{bar}
}
\description{
Testing logging INFO from package
}
